+++ src/lib/marketing/index.ts
@@ -5,15 +5,17 @@ export { default as Header } from './Header/Header.svelte';
 export { default as Container } from './Container/Container.svelte';
 
 export { default as Docs } from './Docs/Docs.svelte';
-export { default as DocsNav } from './Docs/Nav/Nav.svelte';
-export { default as DocsNavItem } from './Docs/Nav/NavItem.svelte';
-export { default as DocsNavCategory } from './Docs/Nav/NavCategory.svelte';
-export { default as DocsContent } from './Docs/Content/Content.svelte';
-export { default as DocsImage } from './Docs/Content/DocsImage.svelte';
+export { default as DocsImage } from './Docs/DocsImage.svelte';
+export { loadDocsPage } from './Docs/fulldocs.js';
+export type {
+	NavSectionConfig,
+	NavConfig,
+	NavPageConfig,
+	NavFoldingSectionConfig,
+	NavSubSectionConfig
+} from './Docs/types.js';
 
 export { default as Footer } from './Footer/Footer.svelte';
 export { default as FooterLinkList } from './Footer/FooterLinkList.svelte';
 
 export { default as Document } from './Document/Document.svelte';
-
-export { default as track } from './track/track.js';
