+++ src/routes/newdocs/[[slug]]/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import Base from '$lib/components/Base/Base.svelte';
-	import FullDocs from '$lib/marketing/Docs/FullDocs.svelte';
+	import Docs from '$lib/marketing/Docs/Docs.svelte';
 	import Header from '$lib/marketing/Header/Header.svelte';
 	import type { PageProps } from './$types.js';
 
@@ -9,5 +9,5 @@
 
 <Base>
 	<Header />
-	<FullDocs {...data} />
+	<Docs {...data} />
 </Base>
